move themepill coloring to its own file
